First tile click is now guaranteed to not be a bomb or have bomb neigbor
Not sure if this was the *best* way to do it, but it was by far the simplest way I could think of.

Tiles, when clicked, now defer to an "on-click" delegate, which is set on initialization by the board component. The board component wires it up to place all the landmines if they haven't been placed yet, and to avoid placing a landmine on the first tile clicked, or on its neighbors.