Take the "bogosort" out of placing landmines
By that, I mean that generation will no longer be hindered by the ratio of "mined" tiles to "clean" tiles.

I have to flatten the gameboard array, and then I just need to take out all of the tiles we don't want to have bombs. Then, it's just as simple as choosing a random tile from the array!